refactor(evals): extract suite builders and add single all-suite entrypoint
