+++ evals/all/README.md
@@ -1,11 +1,8 @@
 # all
 
-Aggregate eval runner package that executes all suites in a single Vitest + LangSmith run.
+Aggregate eval runner package that executes all suites via one top-level
+`ls.describe` block in `index.test.ts`, so the run lands in one LangSmith
+experiment target.
 
-This package sets:
-
-- `LANGSMITH_EVAL_PROJECT=deepagents-js-all`
-- `LANGSMITH_EVAL_DATASET=deepagents-js-all`
-
-so all suites report into one shared LangSmith experiment target when run through
-`@deepagents/eval-all`.
+Each suite exports a `define...Suite(runner)` function from a separate `suite.ts`
+file. `evals/all/index.test.ts` imports those functions and invokes them directly.
