+++ evals/all/vitest.config.ts
@@ -7,8 +7,7 @@ export default defineConfig({
     testTimeout: 120_000,
     hookTimeout: 60_000,
     teardownTimeout: 60_000,
-    include: ["../*/index.test.ts", "../oolong/datasets/*.test.ts"],
-    exclude: ["../all/**"],
+    include: ["index.test.ts"],
     setupFiles: ["@deepagents/evals/setup"],
     reporters: ["langsmith/vitest/reporter"],
   },
