+++ evals/oolong/datasets/imdb.test.ts
@@ -1,15 +1,13 @@
 import * as ls from "langsmith/vitest";
 import { getDefaultRunner } from "@deepagents/evals";
-import { loadOolongTasksByDataset } from "../load-oolong.js";
-import { makeOolongTests } from "../make-tests.js";
+import { defineOolongImdbSuite, getOolongImdbDatasetName, getOolongImdbDescribeOptions } from "./imdb.suite.js";
 
 const runner = getDefaultRunner();
-const tasks = (await loadOolongTasksByDataset()).get("imdb") ?? [];
 
 ls.describe(
-  process.env.LANGSMITH_EVAL_DATASET ?? runner.name,
+  getOolongImdbDatasetName(runner),
   () => {
-    makeOolongTests(tasks);
+    defineOolongImdbSuite(runner);
   },
-  { projectName: process.env.LANGSMITH_EVAL_PROJECT ?? "deepagents-js-oolong-imdb", upsert: true },
+  getOolongImdbDescribeOptions(runner),
 );
