+++ evals/oolong/datasets/metaphors.test.ts
@@ -1,15 +1,13 @@
 import * as ls from "langsmith/vitest";
 import { getDefaultRunner } from "@deepagents/evals";
-import { loadOolongTasksByDataset } from "../load-oolong.js";
-import { makeOolongTests } from "../make-tests.js";
+import { defineOolongMetaphorsSuite, getOolongMetaphorsDatasetName, getOolongMetaphorsDescribeOptions } from "./metaphors.suite.js";
 
 const runner = getDefaultRunner();
-const tasks = (await loadOolongTasksByDataset()).get("metaphors") ?? [];
 
 ls.describe(
-  process.env.LANGSMITH_EVAL_DATASET ?? runner.name,
+  getOolongMetaphorsDatasetName(runner),
   () => {
-    makeOolongTests(tasks);
+    defineOolongMetaphorsSuite(runner);
   },
-  { projectName: process.env.LANGSMITH_EVAL_PROJECT ?? "deepagents-js-oolong-metaphors", upsert: true },
+  getOolongMetaphorsDescribeOptions(runner),
 );
