+++ evals/oolong/datasets/trec_coarse.test.ts
@@ -1,15 +1,13 @@
 import * as ls from "langsmith/vitest";
 import { getDefaultRunner } from "@deepagents/evals";
-import { loadOolongTasksByDataset } from "../load-oolong.js";
-import { makeOolongTests } from "../make-tests.js";
+import { defineOolongTrecCoarseSuite, getOolongTrecCoarseDatasetName, getOolongTrecCoarseDescribeOptions } from "./trec_coarse.suite.js";
 
 const runner = getDefaultRunner();
-const tasks = (await loadOolongTasksByDataset()).get("trec_coarse") ?? [];
 
 ls.describe(
-  process.env.LANGSMITH_EVAL_DATASET ?? runner.name,
+  getOolongTrecCoarseDatasetName(runner),
   () => {
-    makeOolongTests(tasks);
+    defineOolongTrecCoarseSuite(runner);
   },
-  { projectName: process.env.LANGSMITH_EVAL_PROJECT ?? "deepagents-js-oolong-trec-coarse", upsert: true },
+  getOolongTrecCoarseDescribeOptions(runner),
 );
